팔로우 DTO 수정
