--- src/main/java/com/dot/tour_info_service_server/dto/request/follow/test.txt
@@ -1,2 +1,0 @@
-폴더 생성용 txt 입니다.
-dto 생성후 삭제해 주세요.
